requirements.txt
streamlit>=1.28.0
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
torch>=2.0.0
torchvision>=0.15.0
pyyaml>=6.0
albumentations==1.3.1
mediapipe==0.10.21
omegaconf==2.3.0
torchmetrics>=1.0.0
onnx==1.14.1
onnxruntime==1.15.1
opencv-contrib-python==4.8.0.76
pandas==2.0.3
pycocotools==2.0.6
tqdm==4.66.1
